+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # ChangeLog
 
-* fix crash while the multiblock structure width is not equal to the depth.
+* fix gtm energy trait crash
+* fix botania mana trait crash